read: add HDF5 reader based on .h5 extension
An HDF5 file is a hierarchical container format. In order read from it one does
not need only the filename but also a path describing the dataset location
within that dataset. For this, this change adds the "dataset" property which is
used for this purpose but meaningless for the other readers.