write: separate formats into distinct classes
Similar to the general Reader interface, there is now a Writer interface that
can be implemented for various data formats. This change also removes the
"single-file" property which is now guessed from the filename itself. If a
single format specifier is given, we assume to write multiple files, otherwise a
single file with multiple images.