stack: force nonfactor generation
If the number of images in the stack hadn't been a divisor of the number
of input images, the last images would have been dropped. This makes
sure all images are processed and fills the last positions of the stack
(when the total number of stacked images starts to exceed the number of
input images) with arbitrary previous images. Fixes #210.